+++ packages/core/src/utils/get-base-url/index.ts
@@ -1,2 +1,1 @@
 export { getAPIBaseUrl, getMobileScheme, getWebBaseUrl } from "./get-base-url";
-export { validateRedirectUri } from "./validate-redirect-uri";
