--- packages/core/src/utils/getBaseUrl/index.ts
@@ -1,1 +1,0 @@
-export { getAPIBaseUrl, getMobileScheme, getWebBaseUrl } from "./getBaseUrl";
