+++ packages/core/src/utils/validate-redirect-uri/validate-redirect-uri.ts
