--- packages/core/src/utils/validateRedirectUri/index.ts
@@ -1,1 +1,0 @@
-export { validateRedirectUri } from "./validateRedirectUri";
